extract page header into its own component
